// dragonfly/src/app/dragonfly-mod/models/dragonfly.interface.ts
export interface Dragonfly {
    knownAs: string;
    species: string;
    class: string;
    order: string;
    family: string;
    domain: string;
    kingdom: string;
    phylum: string;
    about: string;
    size: string;
    habitat: string;
    diet: string;
    lifeCycle: string;
    lifeSpan: string;
    range: string;
}